+++ src/wai/common/math/__init__.py
@@ -1,0 +1,4 @@
+"""
+Package of mathematical functions and constructs.
+"""
+from ._factorial import factorial