+++ src/wai/common/math/sets/__init__.py
@@ -1,0 +1,7 @@
+"""
+Package for working with sets, subsets and selection.
+"""
+from ._number_of_permutations import number_of_permutations
+from ._number_of_subsets import number_of_subsets
+from ._subset_number_to_subset import subset_number_to_subset
+from ._subset_to_subset_number import subset_to_subset_number